Update facility status reports UX
